requirements.txt
streamlit==1.24.1
streamlit-folium==0.13.0
plotly==5.15.0
numpy==1.24.4
folium==0.14.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
haversine==2.8.0
Pillow==9.5.0